+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Migration step added to delete old `repo` directory. Run `recyclarr migrate` to use.
+
 ## [5.0.3] - 2023-06-25
 
 ### Fixed